# Implement calls to XSMM for all Stripe types. Perf improvements.